% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_questions.R
\name{generate_questions}
\alias{generate_questions}
\title{generate_questions}
\usage{
generate_questions(all_funs, n, theme)
}
\arguments{
\item{all_funs}{function names and their package}

\item{n}{number of questions to generate}

\item{theme}{theme of the functions}
}
\description{
A function to generate a random sample of questions composed of a function name, its package, and
some wrong packages.
}
