% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_options.R
\name{get_golem_options}
\alias{get_golem_options}
\title{get_golem_options}
\usage{
get_golem_options(which = NULL)
}
\arguments{
\item{which}{number or name of element}
}
\description{
get_golem_options
}
