% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hello.R
\name{mod_hello_ui}
\alias{mod_hello_ui}
\alias{mod_hello_server}
\title{mod_hello_ui and mod_hello_server}
\usage{
mod_hello_ui(id)

mod_hello_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module. A "hello" popup to introduce the game and its rules.
}
\keyword{internal}
