% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_playagain.R
\name{mod_playagain_ui}
\alias{mod_playagain_ui}
\alias{mod_playagain_server}
\title{mod_playagain_ui and mod_playagain_server}
\usage{
mod_playagain_ui(id)

mod_playagain_server(input, output, session, timer, score, theme)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{timer}{timer}

\item{score}{current score and number of answered questions}

\item{theme}{theme}
}
\description{
A shiny Module.
}
\keyword{internal}
