% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_question.R
\name{mod_question_ui}
\alias{mod_question_ui}
\alias{mod_question_server}
\title{mod_question_ui and mod_question_server}
\usage{
mod_question_ui(id)

mod_question_server(input, output, session, question, placeholder, event,
  delai)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{question}{data.frame of one row containing a question}
}
\description{
A shiny Module.
}
\keyword{internal}
