% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_realtime_score.R
\name{mod_realtime_score_ui}
\alias{mod_realtime_score_ui}
\alias{mod_realtime_score_server}
\title{mod_realtime_score_ui and mod_realtime_score_server}
\usage{
mod_realtime_score_ui(id)

mod_realtime_score_server(input, output, session, results_question)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{results_question}{output of the question module}
}
\description{
A shiny Module.
}
\keyword{internal}
