% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app_old.R
\name{run_app_old}
\alias{run_app_old}
\title{run_app}
\usage{
run_app_old(usecase, ec_host = "", ec_room = "")
}
\arguments{
\item{usecase}{local, shinyappsio or remote}

\item{ec_host}{ethercalc host}

\item{ec_room}{ethercalc room}
}
\description{
Run the Shiny Application
}
