% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_options.R
\name{with_golem_options}
\alias{with_golem_options}
\title{with_golem_options}
\usage{
with_golem_options(app, golem_opts)
}
\arguments{
\item{app}{shiny app object}

\item{golem_opts}{options}
}
\description{
with_golem_options
}
