% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_user_score.R
\name{mod_add_user_score_ui}
\alias{mod_add_user_score_ui}
\alias{mod_add_user_score_server}
\title{mod_add_user_score_ui and mod_add_user_score_server}
\usage{
mod_add_user_score_ui(id, curr_theme)

mod_add_user_score_server(input, output, session, curr_theme, score,
  usecase = .USECASE, ec_room = .EC_ROOM, ec_host = .EC_HOST)
}
\arguments{
\item{id}{shiny id}

\item{curr_theme}{theme}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{score}{score}

\item{usecase}{usecase}

\item{ec_room}{ethercalc room for saving scores}

\item{ec_host}{ethercalc host}
}
\description{
A shiny Module.
}
\keyword{internal}
